Consolidate {ITEM,SET,SEARCH,AUTOCOMPLETE}_CLICKED
These four actions were already mostly doing the same thing. SET_CLICKED
was dispatched when an item was clicked on the set page, but we also had
code in BuffetHistoryStore that treated ITEM_CLICKED like SET_CLICKED if
the clicked model was an instance of SetObjectModel.

ITEM_CLICKED and SEARCH_CLICKED were already treated exactly the same in
both history stores.

AUTOCOMPLETE_CLICKED was pretty much the same except that it could be
dispatched with a query parameter in the payload as well.

https://phabricator.endlessm.com/T12215