Add Selection.ArticleContext module
This selection will load the articles related to
the one currently being displayed. If the article
was shown as a result of a search, it will show
other articles that match that query. If it was
shown as a result of clicking a set, it will show
other articles in that set.

Makes obselete the searchAndItem layout module.

https://phabricator.endlessm.com/T11737