Don't assume launch comes before window present
The previous code assumed that the dispatcher launch actions, such as
LAUNCHED_FROM_DESKTOP, would come before the
Window.Simple._present_if_needed() call. This is not always the case when
we move this call into make_ready(). This code does the right thing when
the launch action is dispatched later than make_ready().

https://phabricator.endlessm.com/T11266