Move busy cursor functionality to pager
Previously the window would set a busy cursor when various actions were
dispatched, and remove it again when PAGE_READY was dispatched. This gets
rid of the PAGE_READY action. Instead, Pager.Simple sets a busy cursor
when a page transition is requested due to a history store state change,
and removes it when the page transition has completed.

https://phabricator.endlessm.com/T11266