Make sure top title label is visible in template A apps
We have a property binding between the visibility of the
top title label and the visibility of the side title label
(which sits above the table of contents). Only one should
be shown at a time. However, the default, initial state is
for top title to be invisible and side title to be visible.
This will only change if we have a table of contents widget
to show. The problem arises if the first article to get
shown is one that has no table of contents, but needs its
top title to be visible. In this case, the top title
label will never get toggled to become visible.

This fixes that by making the title label visibility match
that of the table of contents in the contructor. This then
propagates to the top title label visibility, courtesy of
the GObject binding.

https://phabricator.endlessm.com/T12065