Remove {SHOW,HIDE}_ARTICLE_SEARCH actions
These are now replaced by a piece of global state, which is for now
parked on the history store. It's a Gio.Action rather than a property so
that we can attach a keybinding to it.

For the keybinding to work, we have to add our history store's action
group to the main window.

https://phabricator.endlessm.com/T12322