preset_b: apply common sidebar styling to all sidebars
We were only applying some styling to the sidebar on the section
and search pages, when really it should be applied to the sidebar
of the article page as well.

The scrollbar on the article page was showing up far darker than it
should because it was not getting picked up by any selectors for this
preset.
https://phabricator.endlessm.com/T12361