Give selections and content groups an error state
Selections now have an 'error' property. When a selection's query fails
for whatever reason, error is set to true, and content groups react to
that by displaying an error message. The actual exception can be
retrieved by calling Selection.get_error_info().

This allows us to get rid of the SEARCH_FAILED action, generalizing it to
all database queries instead of just searches.

https://phabricator.endlessm.com/T11266