Use tag_match_all for engine tags
When using tags to select only sets or only
articles, use the tag_match_all property. We
need to do it in some places to keep tag_match_any
for content tags so it makes sense to be consistent
and do it everywhere so that moltres
can better return fake data.

https://phabricator.endlessm.com/T12388