Check that we have a current model in Filter.Thematic
This filter can get invoked on app start up
(by Selection.AllSets) even though there is no
current model in the history store yet. Obviously
in this case, it can't filter using the current model
so in that situation we just return true in the
include method. Clearly this means the filter
is not actually doing its job but it doesnt matter
because the filter will get invoked again before
the user actually sees this content.

https://phabricator.endlessm.com/T12388