Added FormattableLabel
The FormattableLabel extends Gtk.Label and adds a `text-transform` style
property. With this one, we can transform the label's text through CSS,
using the custom property `-EknFormattableLabel-text-transform`.

Possible values are:
 * 'none' - No transformation is performed. Default.
 * 'uppercase' - The text is transformed to uppercase.
 * 'lowercase' - The text is transformed to lowercase.

Notice that these values need to be specified as strings in the SCSS side of
things for these to be properly interpreted.

https://phabricator.endlessm.com/T11445