theme: default search box theme
We don't have any default theming for search bars that are not in
the top bar at the moment, and two compete appearances for those
in the encyclopedia and mesh themes.

The mesh search bar will now be used across the board as the search
bar theme when not in the top bar, and the encyclopedia theme (which
was not very usable) will go away.
https://phabricator.endlessm.com/T12487