Change how apps run under flatpaks
Previously, applications did not have access to their "real" location
on the filesystem where their data was stored, so we needed logic to
"resolve" the flatpak path before passing it to xapian-bridge.

For 3.0, we need to change this, to make the search provider work. This
also means that we can make datadir.get_content_path() return the real
filesystem location, and remove the custom resolution of flatpak paths.

https://phabricator.endlessm.com/T12444