Layout.DynamicBackground: fix no-results case
If there are no models (e.g. results) in the selection
then draw the background using the default color.

Keep the previous color if there is one already, this
is to avoid unncessary color transitions. E.g. when a
selection is temporarily cleaned we don't want to show
the default color.

https://phabricator.endlessm.com/T12580