Check if need to hide search box after window stops animating
The make_ready refactor meant that there is now an
asynchronous delay from when a page is loaded to when
the Pager stack animates it into view. The problem we
had is that window still thought we we were on the old
page when deciding whether or not to show search box in
top bar. Now we do that check after animation has
complete so we know we are doing the check on the
new page, not the old one.

https://phabricator.endlessm.com/T12388