Add spinner beneath ContentGroup content
Rather than having the spinner be a separate
stack page, add it beneath the arrangement
content. That way, when we need to show it,
it doesn't force the card content off screen,
creating a weird flash experience in between
loading states. Now we simply hide or show it
depending on whether we are in the middle of
loading new content or not.

https://phabricator.endlessm.com/T11274