articleHTMLRenderer: Link articles with fragments correctly
When an article contains a link to a fragment within another page, e.g.
linking to "https://foo.com/bar#baz", we need to look up the crosslink
for "https://foo.com/bar" and translate it into "ekn:///foobar#baz".

Doing this with string.contains isn't the best possible way of doing it,
but it works...