Clear arrangement when selection clears
Previously, if a content group's no-results slot was filled, clearing the
selection would cause it to switch to the no-results page but not remove
the models from the arrangement. Now it does both.

This fixes a bug where the uncleared cards were keeping the content
group's size request too large.

https://phabricator.endlessm.com/T12731