Do not word wrap the title in Thumb cards
We can have text of arbitrary length, and the cards offer very limited
screen real estate; we already tried limiting the content to two lines,
but with the capitalization enabled the most space we're ever going to
get is about 20 characters.

This should avoid long titles pushing out the synopsis from the card.