Sort sets by sequence number.
The AllSets query uses only tags, which are
a boolean term. Therefore the result set wont
be affected by the sort field. However, it will
mean that the array we get back from xapian
is sorted by sequence number, which is needed
for the paraguayan constitution app, where, e.g.
we want to show Part I before Part II.

https://phabricator.endlessm.com/T12831