Add CurrentSet Filter
In the set page of the news preset, we want to:

a) display articles that belong to the current
   set and display preview articles for subsets.

b) we don't want to repeat articles between the
   current set and the subset previews.

The existing approach is to use a filter that
references preview articles content groups, but
that doesn't work since we don't really support
multiple modules with the same factory id.

Therefore, the new approach is to use hierarchy
information of the sets to filter articles that
don't belong to the current set but belong to a
subset.

As a result, we don't display repeated articles.

https://phabricator.endlessm.com/T13152