news preset: change search page scroll behavior
As agreed with @joanafilizola, it is OK to make
the search banner "go away" with the scrolling
as it happens with the set page. This allow us
to style these pages more similarly.

Therefore, re-structure the search page so the
Layout.InfiniteScrolling contains both banner
and ContentGroup, just as we do in set page.

https://phabricator.endlessm.com/T12580