Fix search string highlighting
This regressed on a previous commit because we were using
Pango.parse_markup() to get text to capitalize, which stripped all the
<span>s from the string.

This gets the Pango attribute list from Pango.parse_markup() and
reapplies it, as well as correcting a bug with the accel character. It
also replaces the existing test with a more robust one that looks at the
Pango.Layout directly instead of relying on formatting to be set using
Pango markup.

Original fix by @mattdangerw and @rmacqueen.

https://phabricator.endlessm.com/T13220