FormattableLabel: escape only if use_markup
When use_markup is true, a text like "A & B" must be escaped
to "A &amp; B", otherwise the label won't accept it.

But, when use_markup is false, then the same text should not
be escaped, otherwise it would literally display "A &amp; B"
which is wrong.

Also, add specific test for this special case.

https://phabricator.endlessm.com/T13271