ContentGroup: fix stack clip area
Cards box-shadows are not being drawn properly, e.g. at the
home page of travel app. This happens because GtkStack clips
these shadows when drawn outside of its allocation.

https://bugzilla.gnome.org/show_bug.cgi?id=771436.

The real solution is to fix this in Gtk, but for now it's
safer to "manually" fix this by extending the stack clip
with its child clip.

NOTE that this stack ha no transition so we don't need to
worry about the intermadiate frames, and that the clip is
updated each time we switch between the stack children.

https://phabricator.endlessm.com/T13171