card: fix highlight update
There are two cases we need to update the pango markup for highlight
on a card. When the style context changes, and when the highlight
string changes.

The former was broken, we sometime trigger an old callback on style
context change which would update the card with an old highlight string.
This resulted in a bug where old search queries would remain highlighted
on the home page of preset A.

To fix we remove the update_highlight_string virtual function on card
and instead automatically associate an update callback with a label
inside card's set_label_with_highlight method. This callback will be
called after a style context change or highlight string change.

This still feels a bit nasty, but don't want to do anything more invasive
for a minor release.
https://phabricator.endlessm.com/T13356