ArticleContext: sort by sequence number
Depending on the case, this selection can query using
current set or the search query. Regardless of which
case it is, these queries should be the same ones as
the previous pages, otherwise the articles appear in
a different order and can confuse the users.

Sort the "current set" query by sequence number, since
that is what previous seletion ArticlesForSet does.

https://phabricator.endlessm.com/T13274