LogoSystem: add DynamicLogo widget
This widget implements our logo system. A logo can be
composed of an image, text or both. The logo is setup
via widget and css properties.

We call this dynamic because a) the content is set on
runtime, e.g. we can set the text using i18n, b) the
widget can switch between layouts and sizes depending
on the allocation size and geometry.

https://phabricator.endlessm.com/T13437