LogoSystem: add Banner.Dynamic
This banner is almost identical to Banner.App
except that uses a DynamicLogo widget instead
of a ThemeableImage. This module exposes the
logo system to app developers.

This takes the image part of the logo from a
fixed asset uri and the text part from the
title of the app.

https://phabricator.endlessm.com/T13437