ResponsiveMargins: use real minimum height
From gtk_widget_get_preferred_size docs:
"
Note that the natural height of a height-for-width widget will
generally be a smaller size than the minimum height, since the
required height for the natural width is generally smaller than
the required height for the minimum width.
"

Therefore, make sure we are using the real minimum height when
the widget request mode is set to HEIGTH_FOR_WIDTH, otherwise
the allocated height will always be smaller than the min_size
height and the margins will be completely ignored.

https://phabricator.endlessm.com/T13678