application: error if manual overrides file not found
We wan't to allow for an application to ship without an overrides
file (though we aren't doing it at the moment), that app will render
using a stock theme.

However, the same missing file handler was also silently accepting
the case where a manual overrides file had been specified and was
not found. In that case we should error noisily and fail fast, as
falling back to a stock theme was not what the invoker intended.
https://phabricator.endlessm.com/T13538