Revert "Add Layout.PaginatedScrolling"
This reverts commit b0090ccb1689a1242f228f4cfb2994457df82f16.

Our previous approach to paginated scrolling used a
separate ScrolledWindow. This doesn't work for the
same reason that using a separate scrolled window
in Arrangement.ScrollingList didnt work - namely,
when we put it inside a larger scrolled window, you
now have two separate scrolling mechanisms, which is
disorientating. It also makes sense to attach it to the
content group so that we have direct access to the
selection.