For set-backed cards, show number of articles within set.
We were previously just showing the tags associated with
the content object. This only really makes sense for
articles. For sets, we want to show the number of
articles contained within the set. To do that we walk down
the set hierarchy, using engine queries, until we reach
the 'leaf' nodes, which are articles.

https://phabricator.endlessm.com/T13535