Add 'background-size' property to thumb card
All our previous thumbcard images have always been
set to 'cover' the space provided. In the capacitate
app, we want the set icons to instead sit in the center
of the thumb card's image space. To do this, we need
to add a property to control whether an image
is positioned as 'cover' or 'center'.

https://phabricator.endlessm.com/T13535