Only use width to determine whether to apply margins.
In our apps, we always stick content pages in a scrolled
window. It therefore doesnt make sense to potentially
forgo putting margins on the side of content just because
we don't have enough height.

Instead, now we only look at the width when determining
how much margin to apply to content.

https://phabricator.endlessm.com/T13678