Only clear highlight if new card in arrangement was highlighted
Rather than always clearing highlights across an arrangement when
a new card gets highlighted, only clear it if a new card is
going to be highlighted. This allows us to have, for example,
a subset arrangement with a subset highlighted and a article
arrangement with an article highlighted on the same page.

https://phabricator.endlessm.com/T13535