articleObjectModel: simplify table of contents
This removes the treeNode class we were using for the table of
contents. It would have introduced a dependency for the new C
library on gtk which probably doesn't make sense.

This removes a lot of code complexity, but does expose the raw json
metadata structure which is not particularly easy to work with. We
could bring back a more user friendly data structure in the future.
Preferably in something like a GVariant that will more easily fit
in the C library without adding new dependencies.
https://phabricator.endlessm.com/T13418