Revert "EknVfs: added support for multiples domains"
This reverts commit 1b603fd2f45acdbf41d56f4faa6f7a93d31d4802.

The domain part of the EKN URI is considered legacy and should not be
trusted to contain any specific value. Specifically, for Prensa Libre,
it will contain the old 2.6-style domain for compatibility purposes.

As such, unless we rewrite all EKN URIs at runtime, we cannot support
this feature, and it makes sense to have one global set of EKN URIs
and shards which are of the default domain.

https://phabricator.endlessm.com/T14472