domain: kill domain v2
We'll move to a single domain class supporting the new style
subscription apps.

We will leave the get_domain_impl function around which switches
between domain implementations, in case we need to bring back
multiple domains in the future.

There not to much testing left for domain after this change, but
thinks that's just exposing some holes we've had in our test suite.

This is all about to get ported to C anyway, so I'll look at bringing
back some better coverage on the C side.

https://phabricator.endlessm.com/T13418