application: move in update check
We are going to rewrite domain and engine in the C library, but for
the moment will be keeping the downloader in js as there are big
changes planned with ndn.

This means domain should not have a dependency on the downloader.

We can pretty easily shit things around so application runs the
update directly, and clean up our engine/domain API while we are
at it.
https://phabricator.endlessm.com/T13418