engine: consistently handle app id in exposed api
Most of the engine methods only worked with the default app id, but
get_object_by_query would work with any app id to enable global
search. None of this was documented. And the property name default-app-id
implied that other app ids were usable which was not really true.

This commit renames and moves things around so that methods either
work with any app or are clearly named to use the default application.

If a method requires the default app id and it has not been set, we
will throw an error.
https://phabricator.endlessm.com/T13418