LogoSystem: handle random images nicer
We assumed that the image of the logo would have to
comply with the design specs of the logo system. So
when use an image that doesn't comply with the spec
the logo could be drawn partially.

E.g. com.endlessm.teachers_portal_kindergarten.pt_BR

Make sure that the logo is always drawn completely
regardless of the image geometry.

https://phabricator.endlessm.com/T14128