dynamicLogo: obey min-height and min-width
We had a bug where the nat size of the widget could be smaller
than min size, if the value we set via css with min-width and
min-height was bigger than the hardcoded values.

Also, this fixes the issue we weren't able to enforce the min
sizes beyond the harcoded max sizes.

https://phabricator.endlessm.com/T14128