Move get_subscription_entries out of domain
This function doesn't actually need anything from domain except the app
ID. The subscription entries, however, are needed in order to apply
updates, which should be done before initializing a domain.

In order to fix this circular dependency, we move
get_subscription_entries() out of domain and into utils, and don't
initialize the engine or domain before checking for updates.

https://phabricator.endlessm.com/T14643