Apply downloaded update synchronously
There was already a comment saying to do this, so this change actually
does it. Otherwise, you have a race condition between applying the update
and initializing the domain later on, where apply_update() is copying
manifest.json.new to manifest.json, and domain is trying to read it at
the same time.

Nobody is using the async version of the method anymore, so remove it.

https://phabricator.endlessm.com/T14993