ekncontent: remove get_subscription_entries function from domain
This function doesn't actually need anything from domain except the app
ID. The subscription entries, however, are needed in order to apply
updates, which should be done before initializing a domain.

In order to fix this circular dependency, we will make a new utils
function get_subscriptions_json which both the application code
driving update and the domain can call into.

When downloader moves into the ekncontent library, we can remove
this as public API.
https://phabricator.endlessm.com/T14993