Don't restrict selection results to only articles
Since now template A and B content can include
videos, we don't want to restrict content to just
articles.

Note this tentatively should mean that even small
images, which show up inside articles, show up
in search results. In practice, they are not doing
that right now, I think because their titles never
get indexed. But we should add a tag to be able
to distinguish between content we want to show up
in search results and content we dont.

https://phabricator.endlessm.com/T14516