themeableImage: don't break when image is missing
We started using watermark assets on new presets such as course,
escola and library-list. These assets are provided by the apps
and are not included in the library.

The problem is that these presets break if an app don't provide
the watermark asset, e.g. running another app with ka-customizer.

Therefore, make sure that the ThemeableImage widget won't break
if the uri asset is missing.

https://phabricator.endlessm.com/T15255