eknWebview: don't use destroyed license window
The EknWebview doesn't know when the license window gets destroyed,
e.g. when the user closes the window, so it keeps referencing the
destroyed window.

If the user tries to open the license a second time, the EknWebview
will try to use the previously destroyed window, and fail.

Make sure that the EknWebview can handle this, by reseting its
reference when the window gets destroyed.

https://phabricator.endlessm.com/T14945