library-list: Set page can display articles as well
Previously the library-list's set page would display only subsets. This
makes it impossible to use this template with content with only two
levels of hierarchy, or content with articles and subsets mixed in the
same set.

By using Selection.ContentForSet in the set page, we can display articles
there as well as subsets.

https://phabricator.endlessm.com/T15802