Switch to sassc
Since we cannot really bundle Ruby and the Sass gem into the Endless SDK
flatpak run time, we need to use a different Sass parser to turn our
styles into CSS.

The `sassc` binary uses Libsass, a re-implementation of Ruby Sass from
the same developers. Libsass is portable, faster, and more importantly
it's written in C++, which means we can safely bundle it.