encyclopedia: Use dynamic logo
Replace Banner.App with Banner.Dynamic. There was already some CSS in the
encyclopedia theme to cover that case, so we only have to remove the CSS
for .BannerApp.

This is an API break, but it saves the designers a lot of time, and the
risk is small since the Encyclopedia apps are the only apps that use the
Encyclopedia preset.

https://phabricator.endlessm.com/T16259