encyclopedia: Adjust CSS
The logo font is Libre Baskerville in most cases, so set it as the
default here.

We change the minimum sizes of the logos slightly so that they look good
initially at 1920x1080, while making sure they can still size down small
enough to fit on composite.

We add a bit more space below the search box since the dynamic logos are
a bit larger.

https://phabricator.endlessm.com/T16259