dynamicLogo: control max width and height via CSS props
This widget is limited to harcoded values for max width and height
which don't fit in well in many cases. To fight this, in practise,
programmers started overriding the min width and height values at
the cost of having fixed sizes for the logo, which doesn't behave
properly in different screen sizes.

Having these props will allow us to fully control the size of the
logo as a whole and solve the issue mentioned above without having
to force a fixed size when is not wanted.

e.g:

.BannerDynamic__logo {
    -EknDynamicLogo-max-width: 800px;
    -EknDynamicLogo-max-height: 400px;
    min-width: 200px;
    min-height: 50px;
}

https://phabricator.endlessm.com/T14982