contentGroup: clear hightlight
The ContentGroup listens to changes in the history store
in order highlight its own cards, but it doesn't remove
the highlight after the store chaged to a model that is
not contained in its selection.

Because of this, we sometimes see cards getting selected
apparently at random.

For blog template we need this fixed so that we can show
multiple content groups as one menu.

https://phabricator.endlessm.com/T16557