build: Use locally built introspect to generate docs
We don't want a circular dependency where we depend on a version of the
introspect utility already being installed in order to power the Hotdoc
modular framework extension.

Instead, build a second version of the introspect utility called
introspect-local which uses uninstalled versions of the libraries and
GResources, and make the docs build use that.

https://phabricator.endlessm.com/T14833