modules: show categories in Card.List
Previously, we didn't need to display categories in this
card, since we use it in places where it is redundant to
display this information.

Now we need this for the blog template, so add a option
to enable showing the categories. It will be disabled by
default to keep the existing behavior untouched.

https://phabricator.endlessm.com/T16565