modules: Support any content in article panel
For the blog template, we need more than just the regular
next-article navigation card, instead, we want to display
a ContentGroup with related articles.

The Layout.ArticleStack module is limited to displaying
only 1 navigation card so, to achieve what is described
above, we drop the existing nav-card slot and all of the
specific navigation logic in favor of a more generic slot.

We add a new generic slot called nav-content that can take
any type of content.

Port entries of nav-card to nav-content both for the a)
blog template to display related articles and b) for news
and buffet to mimic the next article behavior.

https://phabricator.endlessm.com/T16699