build: Move docs build to subdirectory
Due to the Hotdoc Makefile snippet overriding special rules like "all"
and "clean", the Hotdoc build rules don't respect the order of SUBDIRS.
However, we need ekncontent to be built before the docs. This changes the
Hotdoc build to take place in a subdirectory so that we can make
absolutely sure that everything in ekncontent and the toplevel directory
is built before the Hotdoc rules take effect.

https://phabricator.endlessm.com/T16717