articleHTMLRenderer: provide content metadata
For the blog template we need to display categories at the top
of the article page and regular tags at the bottom. One way of
doing this would be to hardcode this into the content template
but this won't scale as we are required to add more embedded
contents during runtime.

Therefore, this approach is to provide the content metadata as
as an object so the content creators can access and present it
more flexibly.

Content creators can access the metadata as follows:

<script>
window.contentMetadata
</script>

https://phabricator.endlessm.com/T16869