Allow Order and Filter interfaces to modify query
This adds a method to the Order and Filter interfaces,
modify_xapian_query(), which takes an EosKnowledgeContent.QueryObject and
returns a potentially modified copy of it.

For implementations of Order and Filter, there is a
modify_xapian_query_impl() method to implement so that you don't have to
take care of the logic of when to defer to a sub-order or sub-filter.

https://phabricator.endlessm.com/T13175