Selection.Xapian: modify Xapian queries if needed
This uses the new Order.modify_xapian_query() and
Filter.modify_xapian_query() methods to let the Order and Filter
influence the Xapian query.

This replaces the previous workaround of requesting too many items and
hoping that the number requested was big enough to satisfy the original
request after having been ordered and filtered.

The workaround is still necessary if the Order and/or Filter both can't
modify the query, so we add methods to the Order and Filter interfaces
which tell whether an Order or Filter implementation can modify.

https://phabricator.endlessm.com/T13175