tests: Remove Minimal.MinimalModuleWithSlots
This module was specific to testModule; it never created the modules in
its slots unless specifically asked to. As such, it doesn't really live
up to the expectation of a minimal Module implementation. Move it to the
test where it's used so that it's not confusingly available for other
tests to use.

https://phabricator.endlessm.com/T13175