blog: Fix complementary ContentGroup classes and styling
Previously, we had margins applied to content group, so when
we added a background color to it, the background would only
extend to immediate boundaries of the content group, and not
cover the whole bottom of the article page.

Therefore, switch to padding instead so we can have the same
spacing around the group but be able to extend its background
to cover the whole bottom.

Lastly, add a margin-right to prevent the content group from
hiding the scrolled window slider, and update the class names
used for those content groups.

https://phabricator.endlessm.com/T17076