Context widget (SpaceContainer widget) halign = start
This was messing up the context widget and making its halign property change from
Gtk.Align.START to Gtk.Align.CENTER when the card is minimized to a certain size,
especially when used in the Constrained Arrangement. The halign property for
context should be Gtk.Align.START unless otherwise specified.

http://phabricator.endlessm.com/T15932