New function eknc_query_object_get_query_parser_strings
This builds three separate query strings (one for the user query,
one for filters, and one for excluding filters) to pass to
xapian-bridge.

This means we can take advantage of features of Xapian's query parser
which expect to be dealing with a user-entered query string - so far
stopword handling and incremental search have been moved over.