navigation: Only show the back button if we have > 1 page on stack
Previously we were only hiding it when we were on the homepage,
though that assumed that the homepage was the first page on the
stack. If we open an article from a search provider, then we're
going to end up with that being the first page. Also, the user
should be able to press 'back' if they went to the homepage by
clicking on the home button.

https://phabricator.endlessm.com/T17573