containers: Don't implement vfunc_remove()
Widgets can be removed from containers as a consequence of garbage
collection, and it's not possible to call into JS code during garbage
collection, so this just resulted in a lot of warnings dumped to the
console upon application quit.

Instead, we now do this in C, inside Endless.CustomContainer.

https://phabricator.endlessm.com/T18286