js: Record metrics for article open/close events
With these changes, knowledge apps will send a metrics event to the
daemon each time an article is opened or closed. Included in the event
is the article title, content type, and how it was accessed (e.g. via
desktop search, clicking an article link, etc.) The canonical
documentation for the metrics event data will be on the wiki just like
other events.

This was written in a way that preserves the immutability of HistoryItem
objects, which is why entry point information is passed around as it is.

The commit is a combination of code by Matthew Leeds and Philip Chimento

https://phabricator.endlessm.com/T18516