modules: Tweaks to Audio.Card
I don't think we have gone through design revisions
of the media bin widget yet, neither on how derived
modules should look like, but for now we should try
to make it as decent as possible.

Also, this is adding the basic changes that will
allow us later to focus only on design.

https://phabricator.endlessm.com/T16482