Wire the Bridge operations to the DatabaseManager
This is an initial implementation of the various operation inside
EkncXapianBridge to use the database manager instead of using the
actual HTTP bridge.

The cancellability of the operation is minimal; we will need to move the
GCancellable down into the database manager itself.

This is the minimal implementation that does not touch the logic that
much; after this, we're going to clean up the whole architecture to
remove redundancies.