filter: propagate sub-filter signals
Filters use the "filter-changed" signal to inform the selection
that the filtering criteria has changed, so the selection can
make a new query and get updated content.

The problem is that selections only listen to its own sub-module
filter, while a sub-filter which is the filter's sub-module gets
ignored by the selection.

Make sure that sub-filter signal gets propagated upwards so it
can reach the selection. This fixes the issue where having many
Filter.OtherGroup sub-filters won't affect the selection at all.

https://phabricator.endlessm.com/T18871