historyStore: Record metric when item shared
This adds a new metric that records the app ID, EKN ID, item title,
original URI, and which social network the item was shared on. (We
convert the social networks constants to an enum so that each social
network gets a number instead of a string.)

It is intended to either do nothing or include a note in the payload when
the share is cancelled, but until the share dialog is part of
eos-knowledge-lib that remains unimplemented.

https://phabricator.endlessm.com/T18524