Clean up leaks and invalid accesses
The port from xapian-bridge to direct Xapian access means that we have
to deal with the idiosyncracies of the Xapian API unmediated by layers
of libsoup and json-glib. This requires us to be more careful with the
memory being passed around, and that the strings are what Xapian expects
them to be.