Manually parse some GVariant properties
This mitigates a memory corruption issue that causes critical errors to
be printed in the console when using JSON-GLib to convert JSON into
GVariant for array-based properties that GJS does not support.

The memory corruption is still there, and will present itself given
enough random application launches of some of the knowledge apps.

https://phabricator.endlessm.com/T19561