historyStore: Handle activating media from desktop search
This adds the ability to handle activation of video and audio results
(in addition to articles and sets) from desktop search.

The action will go to the article page and display the media in the
document view. It will not open a lightbox even if the media would
normally be displayed in a lightbox due to being embedded in an article.
The context (in case search results are displayed on the side of the
article page, as in B.yaml) will be the desktop search query.

https://phabricator.endlessm.com/T19814