domain: lock for all database access
Our tests suggest that access to  xapian databases
must be locked against multiple threads. Currently,
we mutex lock only inside EkncDatabaseManager, but
we still  do some access to the  database directly
from EkncDomain.

Therefore, move the locks to EkncDomain to make
sure all access is done with locks.

This fixes a lot of random crashes when querying
for content.

https://phabricator.endlessm.com/T19561