Query both original and corrected terms
If we have a series of spelling corrections, we should merge them with
the original query, to get a query encompassing all the potential
results space.

This allows us to remove a lot of cruft from the QueryObject class, and
to streamline its API so that we can ask for the appropriate query
strings depending on the context.

(cherry picked from commit 6d6ee9776aa47064b45c03c0fbe0c8ff35cadb0b)
Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>