modules: Selection.Suggested should work with less articles
The module assumes that the app contains at least 50 articles
but, when there are less than 50, it ends up crafting a query
that won't get any results, e.g:

An app with only 10 articles, depending on the query, it can
craft a query with offset 12, therefore the offset of 12 get
a empty result set.

The simplest fix is to reduce this assumption.

https://phabricator.endlessm.com/T20020