modules: Restore title style class in Default card
When we split cards in two, cards and views, we moved
set_title_label_from_model from utils to the View
interface.

In the utils version, we added the style class Card_title
but we moved, we changed it to View_title, but we didn't
update the default theme to adjust to this behavior.

So the first step to simply restore the Card style class.

https://phabricator.endlessm.com/T20035