theme: restore Views title color
When we split Cards and Views we  wanted to ensure
that Views titles, like ViewDocument, will default
to white, but then we messed with regular Cards.

Then,  when we fixed this for the Cards we messed
with Views. What this really tells us  is that we
need to think of better way to separate Views and
Cards.

For now, adding an explicit selector to cover both
cases should be enough.

https://phabricator.endlessm.com/T20077