modules: Selection.ArticleContext match with search selection
This module is supposed to derive the context of the article,
e.g. if we go to the article page from a search result, then
the context should display the other results of the search.

Make sure the query for the search case actually matches the
query from Selection.Search, by specifying that is looking
for articles.

https://phabricator.endlessm.com/T20445