ekncontent: Refactor "query" to "search terms"
We use "query" ambiguously throughout the code and this is going to cause
conflicts when adding new API to EkncQueryObject. From now on, "query" is
a query object, and "search terms" are what the user types in. (This
should also prevent us from instinctively writing user-facing error
messages talking about "queries", for which "search terms" is slightly
less jargon-y.)

https://phabricator.endlessm.com/T20373