ekncontent: Fix default values of EkncQueryObject properties
We want the 'limit' property to have a default value of MAXUINT, otherwise
a default query object would be useless since it would retrieve 0 results.

We want the string-valued properties to have a default value of NULL to
indicate they are not set. Otherwise it's easy to forget to check for the
empty string and accidentally make a query that doesn't return anything.

It would be nice to give the GStrv-boxed-valued properties a default value
of NULL as well, but that won't work with g_param_spec_boxed().

https://phabricator.endlessm.com/T20373