ekncontent: Change query construction to use XapianQuery
This removes the string manipulation for constructing the tags, IDs,
filter, and filter-out clauses, and uses XapianQuery directly instead.
This requires API in xapian-glib that is new in the 1.8 series.

We still parse the user input with XapianQueryParser but we now do it
individually for each clause (exact title, title, body, corrected body).
This allows using different feature flags for each, as we do not always
want PARTIAL.

It also solves a few existing weirdnesses with the query parsing. For
example, when we added a '*' to the title in the previous code, it
actually was not wildcarding because we were not using the WILDCARD
feature flag, but since we were using PARTIAL it worked out approximately
to what we wanted.

In the previous code we ran PARTIAL over the whole query string which had
some unintended effects, not wildcarding the terms we expected. This is
also now fixed. Note that this *does* necessarily change the behaviour of
our search queries, as reflected in the tests! The only significant
difference is that we are adding a few more wildcards where we didn't
have them before, so hopefully the effect visible to the user is that
more relevant results are returned.

If we need to tweak this behaviour, it should be much more
straightforward to see the effects now.

https://phabricator.endlessm.com/T20373