autobahn: Add support for overrides via variables
These variables are "points of customization" for our
presets. This means that an app developer can take one
of these presets and change specific parts of it.

To define these variables in the preset:

  vars:
    app-banner:
      shortdef: 'Banner.Dynamic(mode: full, halign: center)'

To reference the variable whithin the preset:

  slots:
    top:
      refvar: $app-banner

Or (short form):

  slots:
    top: $app-banner

To override a variable from the app YAML:

  overrides:
    app-banner:
      shortdef: 'Banner.App(halign: center)'

These variables are defined per preset, and it depends on
what makes sense in each case. This feature will require
us to think deeper about how we structure our presets.

https://phabricator.endlessm.com/T14287