introspect: Process 'allowed' and 'optional' slot attributes
These two are attributes that we've been meaning to add for a long time.
They currently don't have any effect, but we would like to start having
them present on slot attributes so that they show up in the documentation
instead of having to remember to mention it in a documentation comment.

https://phabricator.endlessm.com/T11156