HTML renderer: Don't embed <svg> in <img>
An <img> element has a mandatory src attribute. It happens that if there
isn't one, WebKit will try to load src="null", which gets added to the
base URI and tries to load ekn:///12341234blah/null.

This was the cause of the "Operation not supported" error.

We don't actually need an <img> element, we can just embed the <svg>
directly in the <a> element.

https://phabricator.endlessm.com/T21027