autobahn: Add support for custom overrides
There are cases where we want to customize presets
beyond  the customization variables overrides. So,
we need custom overrides as well.

Custom overrides  allow us  to modify any parts of our
presets, and it doesn't require to do anything special
from the preset side, e.g.:

To use preset B and modify one specific part:

   ---
   overrides:
     root.window.content.content.content.home-page.contents.0.contents.1:
       shortdef: 'Navigation.SearchBox(focus-on-map: false)'
   ---
  !import 'B'

The format is identical to variables and both can be
used together but, instead of a variable name, these
points  are  specified with a path composed of every
node from the root node to the target node.

NOTE that "root" becomes a reserved PREFIX to differ
custom overrides from actual variables, side effect.

https://phabricator.endlessm.com/T14289