theme: More Card/View split style fixes
When we split cards into cards and views we renamed
the  ".DocumentCard" CSS class to ".View", while it
should have been called ".ViewDocument".

Because of this the old ".DocumentCard" style apply
to ALL of our cards, adding style regressions.

Renaming the class (and moving to the  proper file)
gets rid of these regressions, with notable results.

https://phabricator.endlessm.com/T20820