modules: Rework Layout.TopMenu hiding logic
Instead of hiding the bar only if the adjustment
passed a certain threshold, use the direction of
the  adjustment and  a threshold relative to the
current position, to determine if it should hide
or show.

This approach solves the previous issue with
wrong thresholds, by design, and also solves
the issue of making users scroll all the way
up to get the menu back.

https://phabricator.endlessm.com/T21528