widgets: Unify InfiniteScrolledWindow checks
Under  specific  combinations of screen resolution,
arrangement geometry and cards per page, the widget
would not display as many results as it should.

e.g, 1920x1080, 4 rows grid and 10 cards per page,
     in com.endlessm.wiki_art.en app.

To cover for this we used a pixel buffer to detect
and trigger more content but, for some reason, we
do not take it into account for the initial state.

Fix it by unifying all checks so the buffer is
actually used in all cases.

https://phabricator.endlessm.com/T15126