modules: Don't reset scroll when opening lightbox
Reseting the scroll when opening or closing the lightbox
is a regression introduced wnen we refactored the stores
as the HistoryStore now tracks the lightbox as navigation
transition.

Fix it  by adding this case as an exception to the reset
logic. Also unify this logic from both scrolling modules
into one scrolling interface.

https://phabricator.endlessm.com/T20580