modules: Don't show Card context if it's empty
Context-related styling is applied to a intermediary container
and not to  the real visible widget, for practical purposes. A
side effect of doing this is that  the styling applies even if
there is no context info, taking space from the card.

Fix it by not showing the container if no info is provided.

http://phabricator.endlessm.com/T21938