widgets.SearchBox: Simplify cursor determination
In the case where the search box was not on screen, we previously threw
an exception and then analyzed its exception message to see if it should
be ignored. Just returning will achieve the same thing.

We can also get away with fewer calls and coordinate translations, using
Gtk.Entry.get_icon_at_pos().

Also adds a return value of Gdk.EVENT_PROPAGATE to the handler for
clarity.

https://phabricator.endlessm.com/T20353