Support custom modules
A custom module is module that a) is not included
in the SDK, and instead b) it is provided with an
app. These modules are  written by app developers
in Gjs, just like regular modules.

To make sure a custom module will be loaded, the
app developer must add the Gjs source to the app
gresource file, under the /app/js prefix, e.g.:

  <gresource prefix="/app/js">
    <file>custom/custom.js</file>
  </gresource>

Then, to use a custom module it must be called
just like regular modules in the YAML file:

  root:
    type: Custom.Custom

https://phabricator.endlessm.com/T22630