Navigation.SearchBox: Rewrite popup and retheme
This removes the default GtkEntry autocomplete popup in favour of a more
customizable GtkPopover, since a lot of the requested visuals weren't
possible with the default popup (including rounded corners.)

Separates the theming of the popup coming from a search box in the title
bar, from the popup coming from the main search box. Adds mixins to the
SCSS with the various size parameters to ease tweaking in the future.

In some cases (focus out, click) the popup doesn't have the "popdown"
animation. This is because when the window starts animating, the popup's
animation is slotted in after the window's animation, which makes it
jump to the side of the window as its search box starts animating away.

https://phabricator.endlessm.com/T22184